+++ Bayes/testinf.jl
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
 using Turing
 using Phylo
 using Distributions