+++ examples/example_inference.jl
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
 using BenchmarkTools
 using Phylo
 using DataFrames