+++ test/test_show.jl
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
 module TestShow
 
 using Phylo