+++ NEWS.md
@@ -1,5 +1,14 @@
 # NEWS
 
+- v0.5.4
+  - Use ResearchSoftwareMetadata package
+- v0.5.3
+  - Add in metadata and crosswalk
+- v0.5.2
+  - Bugfix in displaying node names
+  - Link fix
+  - Improve testing
+  - Standardise formatting
 - v0.5.1
   - Allow export of trees to nexus and newick format
     - Overload Base.write() and Base.parse()